services/system/requirements.txt
fastapi==0.100.0-beta2
pydantic>=2.0b3
pydantic-settings>=2.0

firebase_admin==5.1.0
google-cloud-logging==2.7.0
google-cloud-pubsub==2.9.0
python-dateutil==2.8.2
pytz==2021.3
ratelimiter==1.2.0.post0
requests==2.26.0
starlette-context==0.3.3
StriveLogger==1.0.*
uvicorn==0.15.0
